+++ mne/epochs.py
@@ -933,22 +933,22 @@ class BaseEpochs(ProjMixin, ContainsMixin, UpdateChannelsMixin,
                            noise_cov=noise_cov, butterfly=butterfly)
 
     @copy_function_doc_to_method_doc(plot_epochs_psd)
-    def plot_psd(self, fmin=0, fmax=np.inf, tmin=None, tmax=None, proj=False,
-                 bandwidth=None, adaptive=False, low_bias=True,
+    def plot_psd(self, fmin=0, fmax=np.inf, tmin=None, tmax=None,
+                 proj=False, bandwidth=None, adaptive=False, low_bias=True,
                  normalization='length', picks=None, ax=None, color='black',
-                 area_mode='std', area_alpha=0.33, dB=True, n_jobs=1,
-                 show=True, average=False, spatial_colors=True,
-                 line_alpha=None, xscale='linear', verbose=None):
+                 xscale='linear', area_mode='std', area_alpha=0.33,
+                 dB=True, estimate='auto', show=True, n_jobs=1,
+                 average=False, line_alpha=None, spatial_colors=True,
+                 verbose=None):
         return plot_epochs_psd(self, fmin=fmin, fmax=fmax, tmin=tmin,
                                tmax=tmax, proj=proj, bandwidth=bandwidth,
                                adaptive=adaptive, low_bias=low_bias,
                                normalization=normalization, picks=picks, ax=ax,
-                               color=color, area_mode=area_mode,
-                               area_alpha=area_alpha, dB=dB, n_jobs=n_jobs,
-                               show=show, average=average,
-                               spatial_colors=spatial_colors,
+                               color=color, xscale=xscale, area_mode=area_mode,
+                               area_alpha=area_alpha, dB=dB, estimate=estimate,
+                               show=show, n_jobs=n_jobs, average=average,
                                line_alpha=line_alpha,
-                               xscale=xscale, verbose=verbose)
+                               spatial_colors=spatial_colors, verbose=verbose)
 
     @copy_function_doc_to_method_doc(plot_epochs_psd_topomap)
     def plot_psd_topomap(self, bands=None, vmin=None, vmax=None, tmin=None,