+++ doc/whats_new.rst
@@ -32,6 +32,8 @@ Changelog
 
 - Add support for showing head surface (to visualize digitization fit) while showing a single-layer BEM to :func:`mne.viz.plot_alignment` by `Eric Larson`_
 
+- plot_psd for Raw and Epochs are refactored into _plot_psd, documentation unified where possible. Both now default to average=False, i.e. coloured channel butterfly plots by `Jeff Hanna`
+
 Bug
 ~~~